Re-added NTAbaddon.dbl from sirgijoe (thanks sirgijoe!)
Addition made to universal config file:
NTConfig_ClearAbaddon = true;			//	Set to true to clear Abaddon of monsters (NOTE: must be running 'Abaddon' in your script array)

When running NTAbaddon.dbl this will set whether or not to clear all of Abaddon.  (Note: This is optional! You have the choice to pop the chest only)
If NTConfig_OpenChests is set to true, bot will open the super-chest in Abaddon.


git-svn-id: https://subversion.assembla.com/svn/d2bs@1294 080cfd25-d5d9-4c59-82ea-70d02a200d8b