This code is KNOWN BROKEN. It CRASHES after the first event occurs.
Reworked the event code to make use of boost and to add the event module to handle events from. Still todo: move everything possible over to boost code.
Also changed the script shutdown behavior to throw an exception on the running script, wait a second, and then consider the script dead.

git-svn-id: https://subversion.assembla.com/svn/d2bs@1311 080cfd25-d5d9-4c59-82ea-70d02a200d8b