Move much of the redundant code for security checks and opening files into a single function in File.cpp.
Files accessible to JavaScript code should use fileOpenRelScript. This function adds the Var.szScriptPath part and does a security check by checking that GetFullPathName of the file to be opened has as a prefix of the GetFullPathName of Var.szScriptPath with a \ added if GetFullPathName does not add it (which it seems not to).

git-svn-id: https://subversion.assembla.com/svn/d2bs@1719 080cfd25-d5d9-4c59-82ea-70d02a200d8b